Optimizer

+++ Project/StaticAssets/ProductionUnits/ProductionUnit.cs
@@ -6,7 +6,7 @@ namespace HeatItOn
         public string ImagePath { get; set; }
         public double MaxHeat { get; set; }
         public double MaxElectricity { get; set; }
-        public int ProductionCosts { get; set; }
+        public double ProductionCosts { get; set; }
         public double CO2Emissions { get; set; }
         public double GasConsumption { get; set; }
     }
